% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor the progress of a script}
\usage{
monitor(
  infile = NA,
  name = NA,
  token = NA,
  API_URL = "https://monitaur-api.herokuapp.com",
  comment_syntax = "#<"
)
}
\arguments{
\item{infile}{Path to the input file}

\item{name}{Name of job (defaults to infile)}

\item{token}{Token for grouping jobs (defaults to uuid)}

\item{API_URL}{Base url of the remote server}

\item{comment_syntax}{Special comment syntax (defaults to #<)}
}
\value{
The result of the script.
}
\description{
This function is used to monitor the progress of a script. A script file
passed to the function will be evaluated and any comments with the
\code{comment_syntax} #< (default) will be sent to a remote sever (\code{API_URL}) for
logging. Before the script is evaluated the job will be initialised with the
remote server and a Job ID is generated. That Job ID is used to link each
step with the script. Once the script is finished the Job satus will be set to
complete. If the script errors the error will also be logged with the server.
}
